requirements.txt
httpx
pandas
sqlalchemy
python-dotenv
psycopg2-binary
streamlit
july
matplotlib~=3.9.1
